requirements.txt
Pillow
pydantic
numpy
opencv-python-headless
alibabacloud_imageseg20191230
alibabacloud_tea_openapi
alibabacloud_tea_util
fastapi
uvicorn
python-multipart
httpx
fal-client
